term: keymapforce command for lack of at_keyboard
This adds a command named `keymapforce`. Current behaviour is unchanged
until it is run. The use case is people with non-QWERTY keyboards on
computers without support for anything else than `terminal_input
console`. For example, as a Dvorak user on a Microsoft Surface Pro 7,
I would like to enter my encryption key at boot in my preferred layout
using the UEFI touchscreen keyboard in order to log in when encryption
is set up.

The new `keymapforce` command is modeled after the `terminal_x`
commands without arguments. This feature is disabled by default to
avoid regressions, and this command toggles it. A
`grub_term_force_keymap` function pointer global variable is used to
hook the end of the key reception logic. The new code simulates reverse
engineering the intercepted key back into its source on a US keyboard
layout, and then replaying the calculated key code on the keyboard with
the proper layout. This works around any specific `terminal_input`s
that don't support key mapping and insist on passing the character
instead of the key code. This way of moving the key remapping to until
after the specific terminal handlers was the only viable way of adding
support for gettings key mapping working on `terminal_input console`.

I wanted to avoid the global hook by adding instead another
`terminal_input` type but that would complicate fini and memory safety.
I tried to put the code in `console.c` but got linker errors and
software architectural obstacles. The reverse lookup is a loop but
it's still fast, close to o(1), and what I wanted to do being
calculating the inverse of the `layout_us` map manually would be hard
to maintain to keep in sync without `static_assert`.

This bug has existed for 10 years. Many users are affected by this bug:
- bbs.archlinux.org/viewtopic.php?id=240739
- archived.forum.manjaro.org
/t/does-luks-in-manjaro-support-azerty-keyboards/87582/6
- bugs.launchpad.net/ubuntu/+source/grub2/+bug/1914953
- www.mail-archive.com
/search?l=bug-grub@gnu.org&q=subject:%22RE%5C%3A
+Changing+default+keyboard+layout%22&o=newest&f=1
- askubuntu.com/a/1391658/1004020

I have tested the following:
- Machines
  - VirtualBox BIOS Ubuntu 22.04
  - surface-linux Ubuntu 22.10
    - UEFI touchscreen
    - Type cover keyboard
- Actions
  - Timeout then boot
  - Enter to boot normally
  - `c` to command line then running a few commands
- `terminal_input`
  - `console`
  - `at_keyboard`
- `terminal_output`
  - `gfxterm`
  - `console`
- `keymap` from `grub-mklayout`
  - `none` (qwerty)
  - `dvorak`
  - `us -variant colemak`
  - `fr` (azerty)
    - The accents don't work but the ASCII does. Someone can add
Unicode support in another patch.

Signed-off-by: Daniel Tang <danielzgtg.opensource@gmail.com>